Resolv and sane site file
This adds a very basic resolv class to kick things off and make sure
I have hiera set up in some kind of sane way. This also adds in a site
file that include base -> site -> nodeclass in order based on the info
that is gathers from collins at the start of it's run.