Add firstboot script
This adds a firstboot script for getting your system into a desired
state the first time that it comes up. params.pp had to be moved into
functions because it was in the wrong place but magically works as long
as you don't user `puppet apply -e` with the module.