Checkpoint on plex
Going down a rabbit hole... Checking in what I have so far and changing
the OS to 18.04 so I don't have to deal with the python 3.5 vs python 3.6
bs when install pipenv. It does support older version of python but the
ppa that build the package does not ppa:pypa/ppa.