Order classes
Order classes in manifest to try and keep logic super specific to
my environment out of the modules... Although in order to use this
right now you have to have collins setup and run bionic so I don't
think this will help anyone. Will try and clean more of this up
later so it's no so tied into my setup.