Move to official plex docker
Found an official plex docker container that is a little bit nicer
to use. This is a little messy because plex makes initial server setup
a pain in the ass. I don't feel like digging through the sqlite db to
find out what is going on so i left some docs how to do the initial
setup.