Fix collins.json file perms
This isn't a perfect solution but start with making sure that everyone
on the system is not able to cat the file and then drop to root.
Possibly we should just expect the password to be set properly at
provision time and not worry about giving it to puppet for future runs.